:recycle: Move all externalLinks into a const file
