+++ src/components/accountInitialization/index.js
@@ -7,6 +7,7 @@ import { FontIcon } from '../fontIcon';
 import styles from './accountInit.css';
 import fees from './../../constants/fees';
 import { fromRawLsk } from '../../utils/lsk';
+import links from './../../constants/externalLinks';
 
 class AccountInitialization extends React.Component {
   closeInfo() {
@@ -34,7 +35,7 @@ class AccountInitialization extends React.Component {
         <p>{t('It is recommended that you initialize your Lisk ID.')}</p>
         <p>{t('The easiest way to do this is to send LSK to yourself. It will cost you only the usual {{fee}} LSK transaction fee.', { fee: fromRawLsk(fees.send) })}</p>
         <p>
-          <a target='_blank' href='https://help.lisk.io/account-security/should-i-initialize-my-lisk-account' rel='noopener noreferrer'>
+          <a target='_blank' href={links.accountInitialization} rel='noopener noreferrer'>
             {t('Learn more about Lisk ID initialization')} <FontIcon>arrow-right</FontIcon>
           </a>
         </p>
