+++ src/constants/externalLinks.js
@@ -1,0 +1,19 @@
+const links = {
+  helpCenter: 'https://lisk.io/help-center',
+  explainBlockchain: 'https://lisk.io/academy/blockchain-basics/what-is-blockchain/eli5-blockchain',
+  isBlockchainSecure: 'https://lisk.io/academy/blockchain-basics/benefits-of-blockchain/blockchain-security',
+  seeMore: 'https://lisk.io/academy/blockchain-basics/how-does-blockchain-work',
+  liskChat: 'https://lisk.chat/home',
+  helpDeskChannel: 'https://lisk.chat/channel/help-desk',
+  howToStorePassphrase: 'https://lisk.io/academy/blockchain-business/cryptocurrencies/how-to-stay-secure-when-trading-cryptocurrencies',
+  gettingStarted: 'https://lisk.io/academy/welcome-to-the-lisk-academy/getting-started-with-lisk',
+  votingOnLisk: 'https://lisk.io/academy/welcome-to-the-lisk-academy/getting-started-with-lisk/voting-on-lisk',
+  blockchainTransparencyExplained: 'https://lisk.io/academy/blockchain-basics/benefits-of-blockchain/blockchain-transparency-explained',
+  liskAcademy: 'https://lisk.io/academy',
+  privacyPolicy: 'https://lisk.io/privacy',
+  sidechains: 'https://lisk.io/help-center/dapp-development/what-are-sidechains',
+  votingAndDelegates: 'https://lisk.io/help-center#voting-and-delegates',
+  accountInitialization: 'https://lisk.io/help-center/account-security/should-i-initialize-my-lisk-account',
+};
+export default links;
+
