--- src/constants/help.js
@@ -1,10 +1,0 @@
-const links = {
-  helpCenter: 'https://help.lisk.io/',
-  explainBlockchain: 'https://lisk.io/academy/blockchain-basics/what-is-blockchain/eli5-blockchain',
-  isBlockchainSecure: 'https://lisk.io/academy/blockchain-basics/benefits-of-blockchain/blockchain-security',
-  seeMore: 'https://lisk.io/academy/blockchain-basics/how-does-blockchain-work',
-  liskChat: 'https://lisk.chat/home',
-  helpDeskChannel: 'https://lisk.chat/channel/help-desk',
-  howToStorePassphrase: 'https://lisk.io/academy/blockchain-business/cryptocurrencies/how-to-stay-secure-when-trading-cryptocurrencies'
-};
-export default links;
